Move check code into its own module
