+++ src/privateer2/backup.py
@@ -1,5 +1,5 @@
 import docker
-from privateer2.keys import check
+from privateer2.check import check
 from privateer2.util import match_value, mounts_str, run_container_with_command
